Package the code
Why? I did so mainly for two reasons related to my development:

- With the new code in `rewrite/`, I either needed to package the code
  or do some ugly path hacks like adding this code to several files:
  ```python
  import sys
  import os
  sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
  ```

- I can use UV which is pretty sweet, it provides a CLI for managing
  dependencies and the venv and resolves dependencies super fast.

But it might also be useful later on, because it enables having project
configs in pyproject.toml, like for testing and linting.

I packaged the code using astral/uv, "An extremely fast Python package and project
manager, written in Rust.", hence the `uv.lock` file.

I needed to add `bt-decode` as an explicit denpendency because it needed
to be a specific pre-release version (according to the UV output when
adding the bittensor dependency).